requirements.txt
ultralytics<=8.3.0
torch<=2.4.1
torchvision<=0.19.1
torchaudio<=2.4.1
faiss-cpu<=1.8.0
hdbscan<=0.8.33
numpy<=1.26.4
opencv-python<=4.10.0.84
scikit-learn<=1.5.1
pyyaml<=6.0.2
pydantic<=2.8.2
fastapi<=0.111.0
uvicorn<=0.30.1
geohash2<=1.1
pandas<=2.2.2
tqdm<=4.66.4
pillow<=10.4.0
